UpdatesPage: add SnapUpdates tab
