+++ lib/store_app/settings/repo_dialog.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:software/l10n/l10n.dart';
-import 'package:software/store_app/updates/updates_model.dart';
+import 'package:software/store_app/updates/package_updates_model.dart';
 import 'package:software/updates_state.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
